core/src/main/java/com/emarsys/core/app/AppLifecycleObserver.kt
package com.emarsys.core.app

import androidx.lifecycle.Lifecycle
import androidx.lifecycle.LifecycleObserver
import androidx.lifecycle.OnLifecycleEvent
import com.emarsys.core.session.Session

class AppLifecycleObserver(private val session: Session) : LifecycleObserver {

    @OnLifecycleEvent(Lifecycle.Event.ON_START)
    fun onEnterForeground() {
        session.startSession()
    }

    @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
    fun onEnterBackground() {
        session.endSession()
    }
}